rename `gpt` to `llm` in integration
